requirements.txt
aiofiles==23.1.0
alembic==1.9.1
bcrypt==4.0.1
fastapi==0.95.0
isort==5.12.0
Jinja2==3.1.2
passlib==1.7.4
psycopg2-binary==2.9.5
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
SQLAlchemy==1.4.46
uvicorn==0.20.0